requirements.txt
streamlit==1.31.0
ultralytics==8.1.0
Pillow==10.2.0
numpy==1.26.0
pandas==2.1.0
matplotlib==3.8.0
scikit-learn==1.3.0
torch==2.2.0
torchvision==0.17.0
